utils/requirements.txt
lxml==6.0.1
requests==2.32.5
fastapi==0.116.1
uvicorn==0.35.0
langchain==0.3.27
langchain-openai==0.3.33
langchain-ollama==0.3.10
sqlalchemy==2.0.43
psycopg2-binary==2.9.10
beautifulsoup4==4.13.4
python-dotenv
yagmail==0.15.293
